requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.21.0
matplotlib>=3.5.0
scipy>=1.7.0

# Simulation and RL
pybullet>=3.2.0
gymnasium>=0.29.0
stable-baselines3>=2.0.0

# Data processing
pandas>=1.3.0
scikit-learn>=1.0.0

# Visualization
seaborn>=0.12.0
plotly>=5.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.62.0
tensorboard>=2.10.0

# Optional (for advanced features)
opencv-python>=4.5.0
pillow>=9.0.0

